# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/shops/{shopId}/work-orders/{workOrderId}/inspection": {
      "post": {
        "operationId": "InspectionsController_open",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenInspectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inspección con hallazgos y fotos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Abre la inspección (DVI) de la OT; si ya existe, la devuelve.",
        "tags": [
          "inspections"
        ]
      },
      "get": {
        "operationId": "InspectionsController_get",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inspección con hallazgos y fotos (URLs firmadas de 1 h).",
        "tags": [
          "inspections"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/inspection/findings": {
      "post": {
        "operationId": "InspectionsController_addFinding",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFindingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inspección actualizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionView"
                }
              }
            }
          },
          "409": {
            "description": "La OT ya está cerrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agrega un hallazgo (área, título, severidad verde/amarillo/rojo).",
        "tags": [
          "inspections"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/inspection/findings/{findingId}": {
      "patch": {
        "operationId": "InspectionsController_updateFinding",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "findingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFindingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita un hallazgo.",
        "tags": [
          "inspections"
        ]
      },
      "delete": {
        "operationId": "InspectionsController_removeFinding",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "findingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Borra un hallazgo y sus fotos (también en R2).",
        "tags": [
          "inspections"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/inspection/findings/{findingId}/photos": {
      "post": {
        "operationId": "InspectionsController_requestPhotoUpload",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "findingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPhotoUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Foto pending + instrucciones de subida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paso 1 de subir una foto: devuelve una URL firmada (15 min) para hacer PUT directo a R2 con los headers indicados.",
        "tags": [
          "inspections"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/inspection/findings/{findingId}/photos/{photoId}/confirm": {
      "post": {
        "operationId": "InspectionsController_confirmPhoto",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "findingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionPhotoView"
                }
              }
            }
          },
          "409": {
            "description": "Todavía no se subió, o no coincide con lo declarado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paso 2: confirma la subida; la API verifica en R2 tamaño y tipo.",
        "tags": [
          "inspections"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/inspection/findings/{findingId}/photos/{photoId}": {
      "delete": {
        "operationId": "InspectionsController_removePhoto",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "findingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Borra una foto (también en R2).",
        "tags": [
          "inspections"
        ]
      }
    },
...
  },
  "info": {
    "title": "MechMate API",
    "description": "Backend del ecosistema MechMate (ERP de talleres + marketplace de repuestos, RD). Dinero en centavos como string (`Cents`); fechas ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OpenInspectionDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          }
        }
      },
      "AddFindingDto": {
        "type": "object",
        "properties": {
          "area": {
            "type": "string",
            "example": "Frenos",
            "maxLength": 60
          },
          "title": {
            "type": "string",
            "example": "Pastillas delanteras al 20%",
            "maxLength": 200
          },
          "severity": {
            "type": "string",
            "enum": [
              "ok",
              "attention",
              "urgent"
            ],
            "example": "urgent",
            "description": "ok = verde, attention = amarillo, urgent = rojo."
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "required": [
          "area",
          "title",
          "severity"
        ]
      },
      "UpdateFindingDto": {
        "type": "object",
        "properties": {
          "area": {
            "type": "string",
            "example": "Frenos",
            "maxLength": 60
          },
          "title": {
            "type": "string",
            "example": "Pastillas delanteras al 20%",
            "maxLength": 200
          },
          "severity": {
            "type": "string",
            "enum": [
              "ok",
              "attention",
              "urgent"
            ],
            "example": "urgent",
            "description": "ok = verde, attention = amarillo, urgent = rojo."
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          }
        }
      },
      "RequestPhotoUploadDto": {
        "type": "object",
        "properties": {
          "content_type": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/heic"
            ],
            "example": "image/jpeg"
          },
          "size_bytes": {
            "type": "number",
            "example": 2450000,
            "maximum": 10485760,
            "description": "Tamaño EXACTO del archivo en bytes (queda firmado)."
          }
        },
        "required": [
          "content_type",
          "size_bytes"
        ]
      },
...
      "InspectionView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "work_order_id": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectionFindingView"
            }
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "work_order_id",
          "notes",
          "findings",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "InspectionFindingView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "area": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectionPhotoView"
            }
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "area",
          "title",
          "severity",
          "notes",
          "photos",
          "created_at"
        ],
        "additionalProperties": false
      },
      "FindingSeverity": {
        "type": "string",
        "enum": [
          "ok",
          "attention",
          "urgent"
        ]
      },
      "InspectionPhotoView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "uploaded"
            ]
          },
          "content_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "number"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "url_expires_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "content_type",
          "size_bytes",
          "url",
          "url_expires_at",
          "created_at"
        ],
        "additionalProperties": false
      },
      "PhotoUploadView": {
        "type": "object",
        "properties": {
          "photo": {
            "$ref": "#/components/schemas/InspectionPhotoView"
          },
          "upload": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string"
              },
              "method": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "headers": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "expires_at": {
                "type": "string"
              }
            },
            "required": [
              "url",
              "method",
              "headers",
              "expires_at"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "photo",
          "upload"
        ],
        "additionalProperties": false
      },
...
    }
  }
}